refactor: update database URL and move metadata to a separate file
