+++ src/app/[locale]/layout.tsx
@@ -1,4 +1,5 @@
-import type {Metadata} from 'next'
+"use server"
+
 import {Inter} from 'next/font/google'
 import './globals.css'
 import {Providers} from '../providers'
@@ -14,27 +15,6 @@ import {notFound} from 'next/navigation'
 import {routing} from '@/i18n/routing'
 
 const inter = Inter({subsets: ['latin']})
-
-export const metadata: Metadata = {
-  title: 'Crypto Exchange App',
-  description: 'Get the latest crypto exchange rates',
-  keywords: ['crypto', 'exchange', 'rates'],
-  openGraph: {
-    title: 'Crypto Exchange App',
-    description: 'Get the latest crypto exchange rates',
-    type: 'website',
-    locale: 'en_US',
-    url: 'https://crypto-exchange-rate.vercel.app',
-    images: [
-      {
-        url: 'https://ccrypto-exchange-rate.vercel.app/cryptocurrent-favicon-black.png',
-        width: 800,
-        height: 600,
-        alt: 'Crypto Exchange App',
-      },
-    ],
-  },
-}
 
 export default async function RootLayout(props: {
   children: React.ReactNode
